+++ packages/tddy-e2e/tests/livekit_terminal_rpc.rs
@@ -89,6 +89,7 @@ mod livekit_tests {
             TerminalServiceServer::new(terminal_service),
             RoomOptions::default(),
             None,
+            None,
         )
         .await?;
         let server_handle = tokio::spawn(async move { server.run().await });
@@ -212,6 +213,7 @@ mod livekit_tests {
             &server_token,
             TerminalServiceServer::new(terminal_service),
             RoomOptions::default(),
+            None,
             None,
         )
         .await?;
@@ -373,6 +375,7 @@ mod livekit_tests {
             tddy_service::TerminalServiceServer::new(terminal_service),
             RoomOptions::default(),
             None,
+            None,
         )
         .await?;
         let server_handle = tokio::spawn(async move { server.run().await });
@@ -498,6 +501,7 @@ mod livekit_tests {
             &server_token,
             TerminalServiceServer::new(terminal_service),
             RoomOptions::default(),
+            None,
             None,
         )
         .await?;